refactor: move mock classes to own project
