+++ Yaapii.Http.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.1000
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33414.496
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Yaapii.Http", "src\Yaapii.Http\Yaapii.Http.csproj", "{0C50CF79-8B01-43DF-B3A6-9E4D791A7225}"
 EndProject
@@ -17,6 +17,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Assets", "Assets", "{168672
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Test.Yaapii.Http", "tests\Test.Yaapii.Http\Test.Yaapii.Http.csproj", "{9CDB0214-FDA6-4C13-B212-BC09AB6F69DC}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Yaapii.Http.Mock", "src\Yaapii.Http.Mock\Yaapii.Http.Mock.csproj", "{2EF83024-1316-4C6E-92D1-A79FF88AFE51}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Test.Yaapii.Http.Mock", "tests\Test.Yaapii.Http.Mock\Test.Yaapii.Http.Mock.csproj", "{70F3994A-A492-46BB-AEB8-625179806362}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -31,6 +35,14 @@ Global
 		{9CDB0214-FDA6-4C13-B212-BC09AB6F69DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9CDB0214-FDA6-4C13-B212-BC09AB6F69DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9CDB0214-FDA6-4C13-B212-BC09AB6F69DC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2EF83024-1316-4C6E-92D1-A79FF88AFE51}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2EF83024-1316-4C6E-92D1-A79FF88AFE51}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2EF83024-1316-4C6E-92D1-A79FF88AFE51}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2EF83024-1316-4C6E-92D1-A79FF88AFE51}.Release|Any CPU.Build.0 = Release|Any CPU
+		{70F3994A-A492-46BB-AEB8-625179806362}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{70F3994A-A492-46BB-AEB8-625179806362}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{70F3994A-A492-46BB-AEB8-625179806362}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{70F3994A-A492-46BB-AEB8-625179806362}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
