+++ src/Yaapii.Http.Mock/AwaitedPort.cs
@@ -1,6 +1,6 @@
 ﻿//MIT License
 
-//Copyright(c) 2020 ICARUS Consulting GmbH
+//Copyright(c) 2023 ICARUS Consulting GmbH
 
 //Permission is hereby granted, free of charge, to any person obtaining a copy
 //of this software and associated documentation files (the "Software"), to deal
@@ -82,7 +82,7 @@ namespace Yaapii.Http.Mock
 
             foreach (var listener in IPGlobalProperties.GetIPGlobalProperties().GetActiveTcpListeners())
             {
-                if (listener.Port == port.Value())
+                if (listener.Port == this.port.Value())
                 {
                     isAvailable = false;
                     break;
