+++ src/Yaapii.Http.Mock/Templates/ITemplate.cs
