+++ src/Yaapii.Http.Mock/Templates/Joined.cs
