+++ src/Yaapii.Http.Mock/Yaapii.Http.Mock.csproj
@@ -1,10 +1,10 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <VersionPrefix>7.0.0</VersionPrefix>
     <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
     <Authors>ICARUS Consulting GmbH</Authors>
-    <Description>Object oriented http client.</Description>
+    <Description>Mock http server based on Yaapii.Http.</Description>
     <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
     <NeutralLanguage>en</NeutralLanguage>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
@@ -24,22 +24,20 @@
       <Visible>false</Visible>
     </Content>
   </ItemGroup>
-  
+
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DocumentationFile>bin\Debug\Yaapii.Http.xml</DocumentationFile>
+    <DocumentationFile>bin\Debug\Yaapii.Http.Mock.xml</DocumentationFile>
     <OutputPath>bin\Debug</OutputPath>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <DocumentationFile>bin\Release\Yaapii.Http.xml</DocumentationFile>
+    <DocumentationFile>bin\Release\Yaapii.Http.Mock.xml</DocumentationFile>
     <OutputPath>bin\Release</OutputPath>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Nito.AsyncEx.Context" Version="5.1.2" />
-    <PackageReference Include="Yaapii.Atoms" Version="3.0.0" />
-    <PackageReference Include="Yaapii.JSON" Version="13.3.2" />
-    <PackageReference Include="Yaapii.Xml" Version="2.0.0" />
+    <PackageReference Include="MockHttpServer" Version="1.4.0" />
+    <ProjectReference Include="..\Yaapii.Http\Yaapii.Http.csproj" />
   </ItemGroup>
 
 </Project>
