+++ tests/Test.Yaapii.Http.Mock/Templates/JoinedTests.cs
@@ -1,6 +1,6 @@
 ﻿//MIT License
 
-//Copyright(c) 2020 ICARUS Consulting GmbH
+//Copyright(c) 2023 ICARUS Consulting GmbH
 
 //Permission is hereby granted, free of charge, to any person obtaining a copy
 //of this software and associated documentation files (the "Software"), to deal
@@ -23,27 +23,29 @@
 using Xunit;
 using Yaapii.Atoms.Text;
 using Yaapii.Http.Fake;
-using Yaapii.Http.Parts;
 using Yaapii.Http.Parts.Bodies;
-using Yaapii.Http.Parts.Headers;
-using Yaapii.Http.Parts.Uri;
 using Yaapii.Http.Requests;
 
 namespace Yaapii.Http.Mock.Templates.Test
 {
-    public sealed class MatchTests
+    public sealed class JoinedTests
     {
         [Fact]
-        public void HasResponse()
+        public void ReturnsPrimaryTemplateResponse()
         {
-            var expected = "expected response";
             Assert.Equal(
-                expected,
+                "correct response",
                 new TextOf(
                     new Body.Of(
-                        new Match(
-                            "path",
-                            new FkWire(expected)
+                        new Joined(
+                            new Conditional(
+                                req => true,
+                                new FkWire("correct response")
+                            ),
+                            new Conditional(
+                                req => true,
+                                new FkWire("wrong response")
+                            )
                         ).Response(
                             new Request()
                         )
@@ -52,38 +54,26 @@ namespace Yaapii.Http.Mock.Templates.Test
             );
         }
 
-        [Fact]
-        public void DoesNotApplyForMissingParts()
+        [Theory]
+        [InlineData(false, false, false)]
+        [InlineData(false, true, false)]
+        [InlineData(false, false, true)]
+        [InlineData(true, true, true)]
+        public void ChecksAllTemplates(bool expected, bool firstTemplateMatches, bool secondTemplateMatches)
         {
-            Assert.False(
-                new Match(
-                    new Method("get"),
-                    new FkWire()
-                ).Applies(
-                    new Request() // has no method part
-                )
-            );
-        }
-
-        [Fact]
-        public void AppliesForMatchingParts()
-        {
-            Assert.True(
-                new Match(
-                    new Parts.Joined(
-                        new Method("get"),
-                        new Path("some/path"),
-                        new QueryParam("someParam", "someValue"),
-                        new BearerTokenAuth("a valid token"),
-                        new Body("important data")
+            Assert.Equal(
+                expected,
+                new Joined(
+                    new Conditional(
+                        req => firstTemplateMatches,
+                        new FkWire()
                     ),
-                    new FkWire()
-                ).Applies(
-                    new Get(
-                        "http://localhost/some/path?someParam=someValue",
-                        new BearerTokenAuth("a valid token"),
-                        new Body("important data")
+                    new Conditional(
+                        req => secondTemplateMatches,
+                        new FkWire()
                     )
+                ).Applies(
+                    new Request()
                 )
             );
         }
